refactor: Move external args parsing into kluctl_project package
