+++ pkg/kluctl_project/external_args.go
@@ -1,4 +1,4 @@
-package deployment
+package kluctl_project
 
 import (
 	"fmt"
